project/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-bitfield==2.2.0
django-environ==0.11.2
django-filter==23.2
django-geo==0.8
django-js-asset==2.1.0
django-mptt==0.14.0
django-restapi==0.1.3
django-separatedvaluesfield==0.6.0
django-timezone-field==6.0
django-uuidfield==0.5.0
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
lxml==4.9.3
packaging==23.1
psycopg2==2.9.7
pyproj==3.6.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sgp4==2.22
six==1.16.0
South==1.0.2
sqlparse==0.4.4
typing_extensions==4.7.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==2.0.4
